requirements.txt
python-dotenv
groq
gradio
anthropic
Pillow
SpeechRecognition
pydub
gtts
elevenlabs